16-step §E flip · stage 1a: widen the lap/column mask UInt8 → UInt16
The part-grid 16-step flip is a staged index-width refactor (byte-identical at
8 cols, per the §E plan). Stage 1a: the LAP / column-loop mask is a bitmask OVER
COLUMNS (bit i = column i), so at 16 columns it needs 16 bits — widen every
carrier UInt8 → UInt16:
  · Router heldColumns / rowHeld[] / process(laneMask:) · Derivations lapColumn()
  · Snapshot rowLaneMask[] · SnapshotBuilder rowLaneResolved · Models rowLane[]?
    (Codable — decode-safe: old ≤255 values load into UInt16) · BuildSceneLogic
    rowLane + Input.staging/perform/playLane · BuildSceneSnapshot.performLane
    (persisted, decode-tolerant UInt16) · Kernel/AU setLaneMask · VC @State
  · BuildSelfTest + tests (lapColumn masks, fuzz lane, rowLane arrays)

Byte-identical at 8 cols (values ≤0xFF are identical in both widths); 1010 green,
iOS builds. Next stages: tap/solo masks → 128-bit, cellSounding → 256, then the
Snap.cols=16 flip + active-width + UI.

+++ AUExtension/AudioUnitViewController.swift
@@ -416,9 +416,9 @@ struct DiagView: View {
     @State var swing = 50
     // MODELESS (2026-07-27): GRID CONTROLS — the verb palette. Radio-armed; INSPECT is functional in 1b, the
     // others render inert until their increments land. EDIT mode survives alongside until verb coverage completes.
-    @State var laneMask: UInt8 = 0     // §5b lap: held column keys (bit i = column i), PERFORM only
-    @State var buildStagingLane: UInt8 = 0   // PER-ROW LAP (Paul 2026-08-19): the BUILD STAGING grid's own column-loop (independent of the play grid)
-    @State var buildPerformLane: UInt8 = 0   // PER-ROW LAP: the BUILD PLAY grid's own column-loop (baked into the composed scene per-row)
+    @State var laneMask: UInt16 = 0     // §5b lap: held column keys (bit i = column i), PERFORM only
+    @State var buildStagingLane: UInt16 = 0   // PER-ROW LAP (Paul 2026-08-19): the BUILD STAGING grid's own column-loop (independent of the play grid)
+    @State var buildPerformLane: UInt16 = 0   // PER-ROW LAP: the BUILD PLAY grid's own column-loop (baked into the composed scene per-row)
     @State var tapAltMask: UInt64 = 0  // §9 item 1 ON TAP (unified ALT): ephemeral per-cell alt flips
     @State var tapMuteMask: UInt64 = 0 // §9 item 1 ON TAP = MUTE: ephemeral per-cell mute
     @State var soloEmitterMask: UInt8 = 0  // §9 item 1 ON TAP = SOLO EMITTERS: the derived emitter solo set
